Added transit_icon.png, this is an icon to show LIGHTRAILS Made some changes to route selection, the stops should now show more accurate text for Lines:

+++ src/johankrig/hotmail/com/DirectionsFragment.java
@@ -190,7 +190,7 @@ public class DirectionsFragment extends Fragment
 	    	TextView footerFare = (TextView) footerView.findViewById(R.id.directionsFare);
 	    	if(route.has("fare"))
 	    	{
-	    		String fareCost = "$"+route.getJSONObject("fare").getInt("value");
+	    		String fareCost = "$"+route.getJSONObject("fare").getDouble("value");
 	    		footerFare.setText(fareCost);
 	    	}
 	     	else
